+++ src/views/httpResponseTextDocumentView.ts
@@ -1,5 +1,14 @@
 import { EOL } from 'os';
-import { languages, Position, Range, TextDocument, ViewColumn, window, workspace } from 'vscode';
+import {
+    languages,
+    Position,
+    Range,
+    TextDocument,
+    TextEditor,
+    ViewColumn,
+    window,
+    workspace,
+} from 'vscode';
 import { SystemSettings } from '../models/configurationSettings';
 import { HttpResponse } from '../models/httpResponse';
 import { PreviewOption } from '../models/previewOption';
@@ -8,7 +17,6 @@ import { formatHeaders } from '../utils/misc';
 import { ResponseFormatUtility } from '../utils/responseFormatUtility';
 
 export class HttpResponseTextDocumentView {
-
     private readonly settings: SystemSettings = SystemSettings.Instance;
 
     protected readonly documents: TextDocument[] = [];
@@ -22,23 +30,32 @@ export class HttpResponseTextDocumentView {
         });
     }
 
-    public async render(response: HttpResponse, column?: ViewColumn) {
+    public async render(response: HttpResponse, column?: ViewColumn): Promise<TextEditor> {
         const content = this.getTextDocumentContent(response);
         const language = this.getVSCodeDocumentLanguageId(response);
         let document: TextDocument;
         if (this.settings.showResponseInDifferentTab || this.documents.length === 0) {
             document = await workspace.openTextDocument({ language, content });
             this.documents.push(document);
-            await window.showTextDocument(document, { viewColumn: column, preserveFocus: !this.settings.previewResponsePanelTakeFocus, preview: false });
+            return window.showTextDocument(document, {
+                viewColumn: column,
+                preserveFocus: !this.settings.previewResponsePanelTakeFocus,
+                preview: false,
+            });
         } else {
             document = this.documents[this.documents.length - 1];
             languages.setTextDocumentLanguage(document, language);
-            const editor = await window.showTextDocument(document, { viewColumn: column, preserveFocus: !this.settings.previewResponsePanelTakeFocus, preview: false });
-            editor.edit(edit => {
+            const editor = await window.showTextDocument(document, {
+                viewColumn: column,
+                preserveFocus: !this.settings.previewResponsePanelTakeFocus,
+                preview: false,
+            });
+            await editor.edit(edit => {
                 const startPosition = new Position(0, 0);
                 const endPosition = document.lineAt(document.lineCount - 1).range.end;
                 edit.replace(new Range(startPosition, endPosition), content);
             });
+            return editor;
         }
     }